+++ packages/react/auto-pages/client-entry.js
@@ -1,3 +1,0 @@
-import { bootstrap } from '../dist/auto_pages_client.mjs'
-
-void bootstrap()