+++ packages/react/src/virtual_faasjs_pages.d.ts
@@ -1,5 +1,0 @@
-declare module 'virtual:faasjs-pages' {
-  import type { PageModules } from './routing'
-
-  export const pageModules: PageModules
-}