+++ templates/react-client/src/greeting/createGreeting.ts
@@ -1,5 +1,0 @@
-export function createGreeting(name: string = 'FaasJS'): { message: string } {
-  return {
-    message: `Hello, ${name}!`,
-  }
-}