+++ templates/react-client/src/pages/index.tsx
@@ -1,6 +1,0 @@
-import App from '../App'
-import { getHomePageState } from '../routing/route-data'
-
-export default function HomePage() {
-  return <App {...getHomePageState()} />
-}